Mayor fixes with images, asets and renders
I fixed some bugs with the image when clicking upload button but not choosing any, this was causing a mayor bugs. Assets created for svg and png files and many others icons changes or upgraded, like female and male no-profile pics image. Added a new  property to user object called "avatarPic" which has the local path to the avatar image png file